Added benchmarks classes for RPC.
